backend/requirements.txt
absl-py==2.2.2
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
audioread==3.0.1
av==10.0.0
babel==2.17.0
bidict==0.23.1
blinker==1.9.0
boto3==1.38.2
botocore==1.38.2
cached_path==1.7.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cn2an==0.5.22
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
ctranslate2==3.24.0
cycler==0.12.1
Cython==3.0.12
dateparser==1.1.8
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.18
Distance==0.1.3
distro==1.9.0
docopt==0.6.2
dtw-python==1.4.4
eng-to-ipa==0.0.2
exceptiongroup==1.2.2
executing==2.2.0
faiss-cpu==1.11.0
fastapi==0.115.12
faster-whisper==0.9.0
ffmpy==0.5.0
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
fugashi==1.3.0
g2p-en==2.1.0
g2pkk==0.1.2
google-api-core==2.25.0rc0
google-auth==2.39.0
google-cloud-core==2.4.3
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gradio==3.48.0
gradio_client==0.6.1
grpcio==1.71.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
inflect==7.0.0
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.2.0
jaconv==0.4.0
jamo==0.4.1
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonlines==1.2.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
langid==1.1.6
librosa==0.9.1
llvmlite==0.43.0
loguru==0.7.2
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mecab-python3==1.0.9
melotts==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
narwhals==1.36.0
nest-asyncio==1.6.0
networkx==2.8.8
nltk==3.9.1
num2words==0.5.12
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
onnxruntime==1.19.2
openai==1.76.0
openai-whisper==20240930
orjson==3.10.16
packaging==25.0
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
plac==1.4.5
platformdirs==4.3.7
pooch==1.8.2
proces==0.1.7
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
pykakasi==2.2.1
pyparsing==3.2.3
pypinyin==0.50.0
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
rich==13.9.4
rpds-py==0.24.0
rsa==4.9.1
s3transfer==0.12.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.11.4
semantic-version==2.10.0
sentence-transformers==4.1.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
stack-data==0.6.3
starlette==0.46.2
sympy==1.13.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
torchaudio==2.7.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
triton==3.3.0
txtsplit==1.0.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
Unidecode==1.3.7
unidic==1.1.0
unidic-lite==1.0.8
urllib3==1.26.20
uvicorn==0.34.2
wasabi==0.10.1
wavmark==0.0.3
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.1.3
whisper-timestamped==1.14.2
wrapt==1.17.2
wsproto==1.2.0
zipp==3.21.0
